chore: Bump Rust version(s)
Upstream base `tree-sitter` has seen some updates this now
incorporates.

It also bumps this crate's version. The scheme chosen is for the
Rust crate's version to be in lockstep with the `tree-sitter`
dependency. Looking at the major tree-sitter libraries,
this seems to be the general style chosen. For example:

- Python: https://github.com/tree-sitter/tree-sitter-python/blob/0dee05ef958ba2eae88d1e65f24b33cad70d4367/Cargo.toml
- TypeScript: https://github.com/tree-sitter/tree-sitter-typescript/blob/198d03553f43a45b92ac5d0ee167db3fec6a6fd6/Cargo.toml

The commit also adjust the Rust code to be
compatible (the `&'static` change was just a lint,
not a compile error). `cargo build` and `cargo
test` now pass.